+++ src/main/java/codes/harsha/co2stat/application/port/in/CollectMetrics.java
@@ -1,6 +1,6 @@
 package codes.harsha.co2stat.application.port.in;
 
 public interface CollectMetrics {
-    double getMaxLast30Days(String sensorId);
-    double getAvgLast30Days(String sensorId);
+    int getMaxLast30Days(String sensorId);
+    int getAvgLast30Days(String sensorId);
 }
