+++ src/main/java/codes/harsha/co2stat/domain/Mesurement.java
@@ -21,7 +21,7 @@ public class Mesurement {
         this.sensor = sensor;
     }
 
-    public double getCo2() {
+    public int getCo2() {
         return co2;
     }
